Change file names, put nodes in one file and change how sections work
