--- src/interfaces/WidgetModel.tsx
@@ -1,8 +1,0 @@
-import { NodeModel } from './nodeModel';
-import { SectionDivision } from './sectionDivision';
-
-export interface WidgetModel extends NodeModel {
-  type: string,
-  options: any,
-  version: number
-}
