--- src/interfaces/nodeModel.tsx
@@ -1,5 +1,0 @@
-export interface NodeModel {
-  type: string,
-  theme?: string,
-  themeOptions?: any
-};
