--- src/interfaces/pageNodeModel.tsx
@@ -1,6 +1,0 @@
-import { NodeModel } from './nodeModel';
-
-export interface pageNodeModel extends NodeModel {
-  children: NodeModel[],
-  sectionDivision?: 'VERTICAL' | 'HORIZONTAL'
-}
