--- src/interfaces/rootNodeModel.tsx
@@ -1,8 +1,0 @@
-import { NodeModel } from './nodeModel';
-import { SectionDivision } from './sectionDivision';
-
-export interface RootNodeModel extends NodeModel {
-  children: NodeModel[],
-  sectionDivision?: SectionDivision,
-  theme: string // TODO, how are we handling themes?
-}
