--- src/interfaces/sectionDivision.tsx
@@ -1,1 +1,0 @@
-export type SectionDivision = 'VERTICAL' | 'HORIZONTAL';
