--- src/interfaces/sectionNodeModel.tsx
@@ -1,8 +1,0 @@
-import { NodeModel } from './nodeModel';
-import { SectionDivision } from './sectionDivision';
-
-export interface SectionNodeModel extends NodeModel {
-  children: NodeModel[],
-  sectionDivision?: SectionDivision,
-  relativeSize: number
-}
