+++ src/interfaces/userModel.ts
@@ -2,5 +2,5 @@ export interface UserModel {
   imgUrl: string,
   email: string,
   name: string,
-  token: string
+  token: string,
 }
