+++ src/interfaces/viewStateModel.ts
@@ -1,5 +1,5 @@
 export interface ViewStateModel {
   editDashboard?: boolean,
   home?: boolean,
-  viewDashboard?: boolean
+  viewDashboard?: boolean,
 }
